requirements.txt
numpy
opencv-python >= 3.0.0
pillow
pyfiglet
scipy
pyglet